\docType{methods}
\name{blockdimension-class}
\alias{blockdimension-class}
\alias{show}
\alias{show,blockdimension-method}
\title{blockdimension class}
\description{
  blockdimension is an S4 class representing the concept of
  block-dimension. It contains two slots: "parts"
  (partitions) and "dims" (dimensions)
}

