\name{get_dim_names}
\alias{get_col_names}
\alias{get_dim_names}
\alias{get_row_names}
\title{Dimension Names}
\usage{
  get_dim_names(x)
}
\arguments{
  \item{x}{a matrix or data frame}
}
\description{
  Get both row and column names \cr Use
  \code{get_dim_names} to extract dimnames \cr Use
  \code{get_row_names} to extract rownames \cr Use
  \code{get_col_names} to extract colnames
}
\keyword{internal}

