\name{incompatible_args}
\alias{incompatible_args}
\title{Non Conformable Arguments}
\usage{
  incompatible_args(operator)
}
\arguments{
  \item{operator}{name of operator}
}
\value{
  stop message
}
\description{
  Stops execution of a binary operator when arguments are
  non-conformable
}
\keyword{internal}

