\name{non_multiblock_args}
\alias{non_multiblock_args}
\title{Multiblock Required}
\usage{
  non_multiblock_args(operator)
}
\arguments{
  \item{operator}{name of operator}
}
\value{
  stop message
}
\description{
  Stops execution of a binary operator when a multiblock
  was expected
}
\keyword{internal}

